Renamed utilities to utils in thinking_dataset
