+++ tests/thinking_dataset/datasets/operations/test_list_datasets.py
@@ -15,7 +15,7 @@ from dotenv import load_dotenv
 from loguru import logger
 from thinking_dataset.tonics.data_tonic import DataTonic
 from thinking_dataset.datasets.operations.list_datasets import ListDatasets
-from thinking_dataset.utilities.text_utils import TextUtils
+from thinking_dataset.utils.text_utils import TextUtils
 
 # Load environment variables from .env file
 load_dotenv()
