+++ thinking_dataset/commands/download.py
@@ -4,10 +4,10 @@
 # @license MIT
 
 import click
-from ..utilities.log import Log
-from ..utilities.exceptions import exceptions
-from ..utilities.logger import logger
-from ..utilities.load_dotenv import dotenv
+from thinking_dataset.utils.log import Log
+from thinking_dataset.utils.exceptions import exceptions
+from thinking_dataset.utils.logger import logger
+from thinking_dataset.utils.load_dotenv import dotenv
 import thinking_dataset.config as config
 from ..tonics.data_tonic import DataTonic
 from ..datasets.dataset import Dataset
