+++ thinking_dataset/commands/export.py
@@ -4,10 +4,10 @@
 # @license MIT
 
 import click
-from ..utilities.log import Log
-from ..utilities.logger import logger
-from ..utilities.load_dotenv import dotenv
-from ..utilities.exceptions import exceptions
+from thinking_dataset.utils.log import Log
+from thinking_dataset.utils.logger import logger
+from thinking_dataset.utils.load_dotenv import dotenv
+from thinking_dataset.utils.exceptions import exceptions
 from ..pipeworks.pipelines.pipeline import Pipeline
