+++ thinking_dataset/commands/prepare.py
@@ -4,11 +4,11 @@
 # @license MIT
 
 import click
-from ..utilities.log import Log
+from thinking_dataset.utils.log import Log
 from ..pipeworks.pipelines.pipeline import Pipeline
-from ..utilities.exceptions import exceptions
-from ..utilities.logger import logger
-from ..utilities.load_dotenv import dotenv
+from thinking_dataset.utils.exceptions import exceptions
+from thinking_dataset.utils.logger import logger
+from thinking_dataset.utils.load_dotenv import dotenv
 
 
 @click.command()
