+++ thinking_dataset/datasets/operations/get_whoami.py
@@ -6,7 +6,7 @@
 """
 
 from .operation import Operation
-from ...utilities.log import Log
+from thinking_dataset.utils.log import Log
 
 
 class GetWhoami(Operation):
