+++ thinking_dataset/datasets/operations/list_datasets.py
@@ -4,7 +4,7 @@
 # @license MIT
 
 from .operation import Operation
-from ...utilities.log import Log
+from thinking_dataset.utils.log import Log
 
 
 class ListDatasets(Operation):
