+++ thinking_dataset/db/database.py
@@ -9,11 +9,11 @@ import pandas as pd
 import thinking_dataset.config as config
 from sqlalchemy import create_engine, exc
 from contextlib import contextmanager
-from ..utilities.execute import execute
+from thinking_dataset.utils.execute import execute
 from .operations.query import Query
 from .operations.fetch import Fetch
 from .database_session import DatabaseSession as Session
-from ..utilities.log import Log
+from thinking_dataset.utils.log import Log
 
 
 class Database:
