+++ thinking_dataset/pipeworks/pipes/add_id_pipe.py
@@ -6,7 +6,7 @@
 import pandas as pd
 import uuid
 from .pipe import Pipe
-from ...utilities.log import Log
+from thinking_dataset.utils.log import Log
 
 
 class AddIdPipe(Pipe):
