+++ thinking_dataset/pipeworks/pipes/handle_missing_values_pipe.py
@@ -5,7 +5,7 @@
 
 import pandas as pd
 from .pipe import Pipe
-from ...utilities.log import Log
+from thinking_dataset.utils.log import Log
 
 
 class HandleMissingValuesPipe(Pipe):
