+++ thinking_dataset/utils/logger.py
@@ -6,7 +6,7 @@
 """
 
 from functools import wraps
-from ..utilities.log import Log
+from thinking_dataset.utils.log import Log
 
 
 def logger(func):
